# restaurant-api-gatway-main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Voting service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/ReservationOrder/delete/{id}": {
            "delete": {
                "description": "Delete page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReservationOrder"
                ],
                "summary": "Delete ReservationOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ReservationOrder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while Deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ReservationOrder/getall": {
            "get": {
                "description": "GetAll page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReservationOrder"
                ],
                "summary": "GetAll ReservationOrder",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "menu_item_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "quantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "reservation_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetAll Successful",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GetAllReservationOrders"
                        }
                    },
                    "401": {
                        "description": "Error while GetAlld",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ReservationOrder/update/{id}": {
            "put": {
                "description": "Update page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReservationOrder"
                ],
                "summary": "Update ReservationOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ReservationOrder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "Update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.ReservationOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/public_vote/create": {
            "post": {
                "description": "Create page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReservationOrder"
                ],
                "summary": "Create Public_Vote",
                "parameters": [
                    {
                        "description": "Create",
                        "name": "Create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.ReservationOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/reservationOrder/getbyid/{id}": {
            "get": {
                "description": "GetById page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ReservationOrder"
                ],
                "summary": "GetById ReservationOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ReservationOrder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetById Successful",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ReservationOrder"
                        }
                    },
                    "401": {
                        "description": "Error while GetByIdd",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restoran/create": {
            "post": {
                "description": "Create page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restoran"
                ],
                "summary": "Create Public_Vote",
                "parameters": [
                    {
                        "description": "Create",
                        "name": "Create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.Restoran"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restoran/delete/{id}": {
            "delete": {
                "description": "Delete page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restoran"
                ],
                "summary": "Delete Restoran",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restoran ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while Deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restoran/getall": {
            "get": {
                "description": "GetAll page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restoran"
                ],
                "summary": "GetAll Restoran",
                "parameters": [
                    {
                        "type": "string",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "phone_number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetAll Successful",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GetAllRestorans"
                        }
                    },
                    "401": {
                        "description": "Error while GetAlld",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restoran/getbyid/{id}": {
            "get": {
                "description": "GetById page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restoran"
                ],
                "summary": "GetById Restoran",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restoran ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetById Successful",
                        "schema": {
                            "$ref": "#/definitions/genprotos.Restoran"
                        }
                    },
                    "401": {
                        "description": "Error while GetByIdd",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restoran/update/{id}": {
            "put": {
                "description": "Update page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restoran"
                ],
                "summary": "Update Restoran",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restoran ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "Update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.Restoran"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "genprotos.GetAllReservationOrders": {
            "type": "object",
            "properties": {
                "reservation_orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genprotos.ReservationOrder"
                    }
                }
            }
        },
...
        "genprotos.GetAllRestorans": {
            "type": "object",
            "properties": {
                "restorans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genprotos.Restoran"
                    }
                }
            }
        },
...
        "genprotos.ReservationOrder": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "menu_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "reservation_id": {
                    "type": "string"
                }
            }
        },
        "genprotos.Restoran": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
    }
}